mysite/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dateparser==0.7.2
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-bootstrap-pagination==1.7.1
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-postman==3.6.2
djangorestframework==3.10.3
djangorestframework-oauth==1.1.0
idna==2.8
inflection==0.3.1
isort==4.3.21
joblib==0.13.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.2
oauthlib==3.1.0
profanity-check==1.0.3
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
